turn regex into utils
