+++ backend/lib/endpoints/organisations.js
@@ -9,6 +9,7 @@ const {
   searchOrganisationsSchema,
   updateOrganisationSchema,
 } = require("./schema/organisations");
+const { searchRegex } = require("../utils");
 
 /*
  * /api/organisations
@@ -79,22 +80,7 @@ async function routes(app) {
 
       // if location is defined, use simple regex text query, in order to use $geoNear
       if (location && keywords) {
-        let cleanKeywords = keywords.replace(/[.*+?^${}()|[\]\\\.]/g, "\\$&");
-        let isLatin = /^[a-zA-Z .*+?^${}()|[\]\\\.]+$/.test(cleanKeywords);
-        const keywordsRegex = new RegExp(
-          cleanKeywords
-            .split(/[ \/,=$%#()-]/gi)
-            .filter((key) => key && key.length > 1)
-            .map((key) =>
-              isLatin && key.length <= 3
-                ? "\\b" + key + "\\b"
-                : isLatin
-                ? "\\b" + key
-                : key,
-            )
-            .join("|") || "\\b\\B",
-          "ig",
-        );
+        const keywordsRegex = searchRegex(keywords)
         filters.push({
           $or: [
             { name: keywordsRegex },
