+++ client/src/components/Feed/Post.js
@@ -25,7 +25,7 @@ import { StyledPostPagePostCard } from "./StyledPostPage";
 import TextAvatar from "components/TextAvatar";
 import { typeToTag } from "assets/data/formToPostMappings";
 import filterOptions from "assets/data/filterOptions";
-import { getOptionText } from "components/Feed/utils";
+import { getOptionText, searchRegex } from "components/Feed/utils";
 import {
   RESET_PAGE,
   NEXT_PAGE,
@@ -65,26 +65,7 @@ const Highlight = ({ text = "", highlight = "" }) => {
   if (!highlight || !highlight.trim()) {
     return text;
   }
-  let cleanKeywords = highlight
-    .replace(/[.*+?^${}()|[\]\\\.]/g, "\\$&")
-    .replace(/\\./g, " ");
-  let isLatin = /^[a-zA-Z .*+?^${}()|[\]\\\.]+$/.test(cleanKeywords);
-  const regex = new RegExp(
-    `(${
-      cleanKeywords
-        .split(/[ \/,=$%#()-]/gi)
-        .filter((key) => key && key.length > 1)
-        .map((key) =>
-          isLatin && key.length <= 3
-            ? "\\b" + key + "\\b"
-            : isLatin
-            ? "\\b" + key
-            : key,
-        )
-        .join("|") || "\\b\\B"
-    })`,
-    "ig",
-  );
+  const regex = searchRegex(highlight);
   const parts = text.split(regex);
   return parts
     .filter((part) => part)
